+++ _P028_BME280.ino
@@ -443,7 +443,7 @@ float Plugin_028_readAltitude(float seaLevel)
 //**************************************************************************/
 // MSL pressure formula
 //**************************************************************************/
-float Plugin_028_pressureElevation(float atmospheric, float altitude) {
+float Plugin_028_pressureElevation(float atmospheric, int altitude) {
   return atmospheric / pow(1.0 - (altitude/44330.0), 5.255);
 }
